+++ AzureDataStudioExtension/CHANGELOG.md
@@ -8,6 +8,14 @@ New features:
 Bug fixes:
 * Fixed connection timeout issue after executing 100 queries against TDS Endpoint*
 
+## [v10.1.1](https://github.com/MarkMpn/Sql4Cds/releases/tag/v10.1.1) - 2025-11-01
+
+Bug fixes:
+* Fixed error when using subquery without alias
+* Fixed sorting by optionset attributes that do not have a corresponding name virtual attribute
+* Fixed using conditional query referencing metadata tables
+* Fixed error when values are returned for unexpected attributes
+
 ## [v10.1.0](https://github.com/MarkMpn/Sql4Cds/releases/tag/v10.1.0) - 2025-08-31
 
 New features: